refactor(web): Talent Network settings as a dedicated page, not a modal
Per live feedback: swap the overlay panel for a plain route at
/my/talent-network (entry button now just links there), swap the emoji
icons for real lucide icons (EyeOff/Globe/VenetianMask), and drop the
public-link-plus-copy card entirely — a "View your public page" link
covers it, no raw URL display or clipboard action needed.
