+++ design-system/scripts/web-token-baseline.json
@@ -72,7 +72,6 @@
   "web/src/lib/components/StatusBoard.svelte": 12,
   "web/src/lib/components/SwipeDeck.svelte": 8,
   "web/src/lib/components/TailorLandingView.svelte": 13,
-  "web/src/lib/components/TalentNetworkPanel.svelte": 1,
   "web/src/lib/components/VerdictView.svelte": 6,
   "web/src/lib/components/VoteControl.svelte": 4,
   "web/src/lib/docs/format.ts": 25,
